manufacturing_consulting/requirements.txt
# Manufacturing Consulting System - Dependencies
# Production-ready requirements for the AI-powered consulting platform

# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# AI/ML
anthropic==0.18.1
openai==1.12.0
tiktoken==0.6.0
sentence-transformers==2.3.1
numpy==1.26.4
scikit-learn==1.4.0

# Document Processing
python-docx==1.1.0
pypdf==4.0.1
reportlab==4.1.0
markdown==3.5.2
jinja2==3.1.3

# Search & Embeddings
chromadb==0.4.22
faiss-cpu==1.7.4

# Task Queue & Background Jobs
celery==5.3.6
redis==5.0.1

# Monitoring & Logging
structlog==24.1.0
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.40.0

# Data Validation
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0

# Utilities
httpx==0.26.0
python-dateutil==2.8.2
pytz==2024.1
tenacity==8.2.3
cachetools==5.3.2

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0
factory-boy==3.3.0

# Development
black==24.1.1
isort==5.13.2
mypy==1.8.0
ruff==0.2.1
